Add support for read-only mode
